requirements.txt
python-telegram-bot==20.3
pandas
numpy
requests
python-dotenv
nest-asyncio